move helpers into common files
