+++ compiler/src/iree/compiler/Codegen/LLVMGPU/LLVMGPUPackSharedMemoryAlloc.cpp
@@ -9,95 +9,16 @@
 #include "iree/compiler/Codegen/LLVMGPU/Utils/LLVMGPUUtils.h"
 #include "iree/compiler/Codegen/PassDetail.h"
 #include "iree/compiler/Codegen/Passes.h"
+#include "iree/compiler/Codegen/Transforms/Transforms.h"
 #include "iree/compiler/Codegen/Utils/GPUUtils.h"
-#include "mlir/Analysis/Liveness.h"
 #include "mlir/Dialect/Arith/IR/Arith.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/GPU/IR/GPUDialect.h"
 #include "mlir/Dialect/NVGPU/IR/NVGPUDialect.h"
 #include "mlir/IR/Dominance.h"
 
-#define DEBUG_TYPE "iree-llvmgpu-pack-shared-memory-alloc"
-
 namespace mlir {
 namespace iree_compiler {
-
-using AliasGroup = SmallVector<Operation *>;
-
-/// Analyze the liverange of allocs and set them in individual groups if
-/// possible and mark allocs with an attributes.
-/// The algorithm is a simplistic memory allocation solution. It sorts
-/// allocations into alias groups. Everytime two alloc's liverange interfers
-/// they are merge into the same group. If a new alloc is part of multiple alias
-/// groups all those are merged into one. At the end we are left with groups of
-/// allocations that are disjoint and can use the same memory.
-// TODO: Move this to a common place if needed by other backends.
-static void analyzeSharedMemoryAlloc(func::FuncOp funcOp,
-                                     const SmallVector<Operation *> &allocs,
-                                     SmallVector<AliasGroup> &aliasGroups) {
-  struct AllocGroup {
-    SmallVector<Operation *> allocs;
-    // Keep track of every operation where any of the alloc in the group is
-    // live.
-    llvm::DenseSet<Operation *> liveness;
-  };
-  Liveness liveness(funcOp);
-  SmallVector<AllocGroup> groups;
-  for (Operation *alloc : allocs) {
-    SmallVector<size_t> aliasGroups;
-    for (size_t i : llvm::seq<size_t>(0, groups.size())) {
-      AllocGroup &group = groups[i];
-      for (Operation *user : alloc->getUsers()) {
-        // Skip the whole analysis if any user is a subview.
-        // TODO: This could be extended if needed by recursively merging
-        // liveness.
-        if (isa<memref::SubViewOp>(user)) return;
-        if (group.liveness.count(user)) {
-          aliasGroups.push_back(i);
-          break;
-        }
-      }
-    }
-    if (aliasGroups.empty()) {
-      // If we didn't find any alias group create a new one.
-      AllocGroup &newGroup = groups.emplace_back();
-      newGroup.allocs.push_back(alloc);
-      Liveness::OperationListT liveInfo =
-          liveness.resolveLiveness(alloc->getResult(0));
-      newGroup.liveness.insert(liveInfo.begin(), liveInfo.end());
-    } else {
-      // Merge the alloc into the first alias group it interfers with.
-      AllocGroup &mergeGroup = groups[aliasGroups[0]];
-      mergeGroup.allocs.push_back(alloc);
-      Liveness::OperationListT liveInfo =
-          liveness.resolveLiveness(alloc->getResult(0));
-      mergeGroup.liveness.insert(liveInfo.begin(), liveInfo.end());
-      // Then merge all the other alias groups into the first group.
-      for (size_t i = 1, e = aliasGroups.size(); i < e; i++) {
-        AllocGroup &group = groups[aliasGroups[i]];
-        mergeGroup.allocs.insert(mergeGroup.allocs.end(), group.allocs.begin(),
-                                 group.allocs.end());
-        mergeGroup.liveness.insert(group.liveness.begin(),
-                                   group.liveness.end());
-        // For simplicity we leave the group empty and don't remove it.
-        group.allocs.clear();
-        group.liveness.clear();
-      }
-    }
-  }
-
-  LLVM_DEBUG({
-    for (size_t i = 0; i < groups.size(); i++) {
-      llvm::dbgs() << "Alias group " << i << ":\n";
-      for (Operation *op : groups[i].allocs) op->dump();
-    }
-  });
-
-  for (size_t i = 0; i < groups.size(); i++) {
-    if (groups[i].allocs.empty()) continue;
-    aliasGroups.push_back(std::move(groups[i].allocs));
-  }
-}
 
 /// Insert barriers and wait operations if there are allocs of a different alias
 /// group before the given alloc.
@@ -137,14 +58,6 @@ static void addBarrier(func::FuncOp funcOp, Operation *alloc,
   builder.create<gpu::BarrierOp>(alloc->getLoc());
 }
 
-static int64_t getAllocSize(Operation *op, DataLayout &dataLayout) {
-  auto allocOp = cast<memref::AllocOp>(op);
-  int64_t numElements = allocOp.getType().getNumElements();
-  return (dataLayout.getTypeSizeInBits(allocOp.getType().getElementType()) *
-          numElements) /
-         8;
-}
-
 namespace {
 
 struct LLVMGPUPackSharedMemoryAllocPass
@@ -181,37 +94,7 @@ struct LLVMGPUPackSharedMemoryAllocPass
     }
 
     OpBuilder builder(funcOp.getContext());
-    DataLayout dataLayout = DataLayout::closest(funcOp);
-    builder.setInsertionPointToStart(&(*funcOp.getBody().begin()));
-    int64_t maxAlloc = 0;
-    for (size_t i = 0; i < aliasGroups.size(); i++) {
-      int64_t allocSize = 0;
-      for (Operation *alloc : aliasGroups[i]) {
-        allocSize += getAllocSize(alloc, dataLayout);
-      }
-      maxAlloc = std::max(maxAlloc, allocSize);
-    }
-
-    auto workgroupSpace = gpu::AddressSpaceAttr::get(
-        builder.getContext(), gpu::GPUDialect::getWorkgroupAddressSpace());
-    MemRefType allocType = MemRefType::get({maxAlloc}, builder.getI8Type(),
-                                           AffineMap(), workgroupSpace);
-    Value packedAlloc =
-        builder.create<memref::AllocOp>(funcOp.getLoc(), allocType);
-    for (size_t i = 0; i < aliasGroups.size(); i++) {
-      int64_t offset = 0;
-      for (Operation *alloc : aliasGroups[i]) {
-        Location loc = alloc->getLoc();
-        builder.setInsertionPoint(alloc);
-        Value offsetValue = builder.create<arith::ConstantIndexOp>(loc, offset);
-        Value newAlloc = builder.create<memref::ViewOp>(
-            packedAlloc.getLoc(), alloc->getResultTypes()[0], packedAlloc,
-            offsetValue, ArrayRef<Value>({}));
-        offset += getAllocSize(alloc, dataLayout);
-        alloc->replaceAllUsesWith(ArrayRef<Value>({newAlloc}));
-        alloc->erase();
-      }
-    }
+    packAllocs(builder, funcOp, aliasGroups);
   }
 };
 }  // namespace
