+++ compiler/src/iree/compiler/Codegen/Transforms/Transforms.h
@@ -96,6 +96,26 @@ void populateReshapeToInterfaceTensorPatterns(RewritePatternSet &patterns);
 /// Populate patterns that remove dead allocations
 void populateRemoveDeadMemAllocPatterns(RewritePatternSet &patterns);
 
+// Group of Alloc operations that have overlapping liveranges.
+using AliasGroup = SmallVector<Operation *>;
+
+/// Analyze the liverange of the given allocs and set them in individual groups
+/// if they don't overlap.
+/// The algorithm is a simplistic memory allocation solution. It sorts
+/// allocations into alias groups. Everytime two alloc's liverange interfers
+/// they are merge into the same group. If a new alloc is part of multiple alias
+/// groups all those are merged into one. At the end we are left with groups of
+/// allocations that are disjoint and can use the same memory.
+void analyzeSharedMemoryAlloc(func::FuncOp funcOp,
+                              const SmallVector<Operation *> &allocs,
+                              SmallVector<AliasGroup> &aliasGroups);
+
+/// Pack groups of allocations into a unique large i8 allocation and use
+/// memref.view to separate the indivudual allocations. This allows re-using
+/// memory across alias groups.
+void packAllocs(OpBuilder &builder, func::FuncOp funcOp,
+                SmallVector<AliasGroup> &aliasGroups);
+
 }  // namespace iree_compiler
 }  // namespace mlir
